ref(onboarding): Split dart onboarding doc into multiple files
