+++ static/app/gettingStartedDocs/dart/dart/index.tsx
@@ -1,0 +1,14 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+
+import {crashReport} from './crashReport';
+import {logs} from './logs';
+import {onboarding} from './onboarding';
+
+const docs: Docs = {
+  onboarding,
+  feedbackOnboardingCrashApi: crashReport,
+  crashReportOnboarding: crashReport,
+  logsOnboarding: logs,
+};
+
+export default docs;
