+++ static/app/gettingStartedDocs/dart/dart/onboarding.spec.tsx
@@ -2,7 +2,7 @@ import {renderWithOnboardingLayout} from 'sentry-test/onboarding/renderWithOnboa
 import {screen} from 'sentry-test/reactTestingLibrary';
 import {textWithMarkupMatcher} from 'sentry-test/utils';
 
-import docs from './dart';
+import docs from '.';
 
 describe('dart onboarding docs', () => {
   it('renders docs correctly', async () => {
